# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Dev Stack Advisor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Budget range for the project
 */
export type ProjectInputBudget =
  (typeof ProjectInputBudget)[keyof typeof ProjectInputBudget];

export const ProjectInputBudget = {
  zero: "zero",
  low: "low",
  medium: "medium",
  high: "high",
  enterprise: "enterprise",
} as const;

/**
 * Time available for the project
 */
export type ProjectInputTimeLimit =
  (typeof ProjectInputTimeLimit)[keyof typeof ProjectInputTimeLimit];

export const ProjectInputTimeLimit = {
  jam: "jam",
  month: "month",
  quarter: "quarter",
  year: "year",
  longterm: "longterm",
} as const;

/**
 * Developer skill level
 */
export type ProjectInputSkillLevel =
  (typeof ProjectInputSkillLevel)[keyof typeof ProjectInputSkillLevel];

export const ProjectInputSkillLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
  expert: "expert",
} as const;

/**
 * Team size
 */
export type ProjectInputTeamSize =
  (typeof ProjectInputTeamSize)[keyof typeof ProjectInputTeamSize];

export const ProjectInputTeamSize = {
  solo: "solo",
  small: "small",
  medium: "medium",
  large: "large",
} as const;

/**
 * Art and design capability level
 */
export type ProjectInputArtCapability =
  (typeof ProjectInputArtCapability)[keyof typeof ProjectInputArtCapability];

export const ProjectInputArtCapability = {
  none: "none",
  basic: "basic",
  intermediate: "intermediate",
  advanced: "advanced",
  professional: "professional",
} as const;

export interface ProjectInput {
  /** Description of the game project idea */
  projectIdea: string;
  /** Budget range for the project */
  budget: ProjectInputBudget;
  /** Time available for the project */
  timeLimit: ProjectInputTimeLimit;
  /** Developer skill level */
  skillLevel: ProjectInputSkillLevel;
  /** Team size */
  teamSize: ProjectInputTeamSize;
  /** Target platforms (pc, mobile, web, console, vr) */
  platformTarget: string[];
  /** Art and design capability level */
  artCapability: ProjectInputArtCapability;
  /**
   * Any other constraints or notes
   * @nullable
   */
  otherConstraints?: string | null;
}

export interface ToolRecommendation {
  toolId: number;
  toolName: string;
  /** Fit score 0-100 */
  score: number;
  /** Why this tool was recommended */
  reasoning: string;
  strengths: string[];
  weaknesses: string[];
  tradeoffs: string;
  isTopPick: boolean;
}

export interface CategoryRecommendation {
  category: string;
  categoryLabel: string;
  topPick: ToolRecommendation;
  alternatives: ToolRecommendation[];
  categoryReasoning: string;
}

export interface AnalysisResult {
  sessionId: number;
  /** AI-generated summary of the project */
  projectSummary: string;
  /** Detected game genre/type */
  detectedProjectType: string;
  categories: CategoryRecommendation[];
  /** Overall confidence score 0-100 */
  overallConfidence: number;
  /** Final narrative explanation of the stack */
  finalSummary: string;
  /** One-liner overview of the recommended stack */
  stackOverview: string;
}

export interface SessionSummary {
  id: number;
  projectIdea: string;
  detectedProjectType: string;
  stackOverview: string;
  overallConfidence: number;
  createdAt: string;
}

export interface Session {
  id: number;
  projectInput: ProjectInput;
  result: AnalysisResult;
  createdAt: string;
}

export type ToolPricing = (typeof ToolPricing)[keyof typeof ToolPricing];

export const ToolPricing = {
  free: "free",
  freemium: "freemium",
  paid: "paid",
  subscription: "subscription",
  open_source: "open_source",
} as const;

export type ToolMinSkillLevel =
  (typeof ToolMinSkillLevel)[keyof typeof ToolMinSkillLevel];

export const ToolMinSkillLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
  expert: "expert",
} as const;

export interface Tool {
  id: number;
  name: string;
  category: string;
  description: string;
  /** @nullable */
  website?: string | null;
  pricing: ToolPricing;
  minSkillLevel: ToolMinSkillLevel;
  platforms: string[];
  strengths: string[];
  weaknesses: string[];
  bestFor: string[];
  tags: string[];
}

export interface ToolCategory {
  id: string;
  label: string;
  description: string;
  toolCount: number;
}

export type AdvisorStatsTopRecommendedToolsItem = {
  toolName: string;
  count: number;
};

export type AdvisorStatsPopularCategoriesItem = {
  category: string;
  count: number;
};

export interface AdvisorStats {
  totalAnalyses: number;
  topRecommendedTools: AdvisorStatsTopRecommendedToolsItem[];
  popularCategories: AdvisorStatsPopularCategoriesItem[];
  avgConfidenceScore: number;
}

export type ListToolsParams = {
  category?: string;
};