# lib/api-zod/src/generated/types/advisorStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Dev Stack Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { AdvisorStatsPopularCategoriesItem } from "./advisorStatsPopularCategoriesItem";
import type { AdvisorStatsTopRecommendedToolsItem } from "./advisorStatsTopRecommendedToolsItem";

export interface AdvisorStats {
  totalAnalyses: number;
  topRecommendedTools: AdvisorStatsTopRecommendedToolsItem[];
  popularCategories: AdvisorStatsPopularCategoriesItem[];
  avgConfidenceScore: number;
}