# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Dev Stack Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryRecommendation } from "./categoryRecommendation";

export interface AnalysisResult {
  sessionId: number;
  /** AI-generated summary of the project */
  projectSummary: string;
  /** Detected game genre/type */
  detectedProjectType: string;
  categories: CategoryRecommendation[];
  /** Overall confidence score 0-100 */
  overallConfidence: number;
  /** Final narrative explanation of the stack */
  finalSummary: string;
  /** One-liner overview of the recommended stack */
  stackOverview: string;
}