# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Dev Stack Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectInputArtCapability } from "./projectInputArtCapability";
import type { ProjectInputBudget } from "./projectInputBudget";
import type { ProjectInputSkillLevel } from "./projectInputSkillLevel";
import type { ProjectInputTeamSize } from "./projectInputTeamSize";
import type { ProjectInputTimeLimit } from "./projectInputTimeLimit";

export interface ProjectInput {
  /** Description of the game project idea */
  projectIdea: string;
  /** Budget range for the project */
  budget: ProjectInputBudget;
  /** Time available for the project */
  timeLimit: ProjectInputTimeLimit;
  /** Developer skill level */
  skillLevel: ProjectInputSkillLevel;
  /** Team size */
  teamSize: ProjectInputTeamSize;
  /** Target platforms (pc, mobile, web, console, vr) */
  platformTarget: string[];
  /** Art and design capability level */
  artCapability: ProjectInputArtCapability;
  /**
   * Any other constraints or notes
   * @nullable
   */
  otherConstraints?: string | null;
}