# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Dev Stack Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from "./analysisResult";
import type { ProjectInput } from "./projectInput";

export interface Session {
  id: number;
  projectInput: ProjectInput;
  result: AnalysisResult;
  createdAt: Date;
}