# lib/api-zod/src/generated/types/sessionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Dev Stack Advisor API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionSummary {
  id: number;
  projectIdea: string;
  detectedProjectType: string;
  stackOverview: string;
  overallConfidence: number;
  createdAt: Date;
}