# lib/api-zod/src/generated/types/toolCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Dev Stack Advisor API
 * OpenAPI spec version: 0.1.0
 */

export interface ToolCategory {
  id: string;
  label: string;
  description: string;
  toolCount: number;
}