# lib/api-zod/src/generated/types/toolRecommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Dev Stack Advisor API
 * OpenAPI spec version: 0.1.0
 */

export interface ToolRecommendation {
  toolId: number;
  toolName: string;
  /** Fit score 0-100 */
  score: number;
  /** Why this tool was recommended */
  reasoning: string;
  strengths: string[];
  weaknesses: string[];
  tradeoffs: string;
  isTopPick: boolean;
}